RSSI: Migrate from copying data into owned slabs
This replaces the previous bump allocator that provided a view on the
internal memory (with a short lifetime that precluded concurrent
operations) with an owned chunk of a buffer.

Its alternative is one more round of copying data.

This paves the way for receiving RSSI data in a more streaming fashion.